requirements.txt
agent-framework-azure-ai==1.0.0b260107
agent-framework-core==1.0.0b260107
azure-identity>=1.17.0
python-dotenv>=1.0.0
streamlit>=1.40.0
